+++ src/main/java/com/global/api/entities/DisputeDocument.java
@@ -6,6 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class DisputeDocument {
+    private String id;
     private String type;
     private String base64Content;
 }