+++ src/test/java/com/global/api/tests/gpapi/GpApiCreditWithMerchantIdTests.java
@@ -24,7 +24,7 @@ import static org.junit.Assert.*;
 public class GpApiCreditWithMerchantIdTests extends BaseGpApiTest {
 
     private CreditCardData card;
-    private final String merchantId = "MER_c4c0df11039c48a9b63701adeaa296c3";
+    private final String merchantId = "MER_7e3e2c7df34f42819b3edee31022ee3f";
 
     @Before
     public void initialize() throws ConfigurationException {